+++ src/main/webapp/WEB-INF/jsp/viewproject.jsp
@@ -31,7 +31,7 @@
   <div id="body">
 
   <a href="<%=request.getContextPath()%>/html/editproject?id=${pw.project.id}">Edit</a> | 
-  <a href="<%=request.getContextPath()%>/html/deleteproject?id=${pw.project.id}">Delete</a>
+  <a onclick="return (confirm('Are you sure you want to delete this project?'))" href="<%=request.getContextPath()%>/html/deleteproject?id=${pw.project.id}">Delete</a>
   
   <h3>Project: ${pw.project.name}</h3>
